Restored drag drop functionality. exported mqtt server and socket function handlers to external files. removed unnecessary functions. fixed frontend to render widgets correctly.
