+++ client/src/QRCode.js
@@ -1,6 +1,6 @@
 import * as React from "react";
 import {generateQRCode} from './api/get';
-import qrcode from './savedQrCode/qrcode.svg';
+//import qrcode from 'http://localhost:5000/public/savedQRCode/qrcode.svg';
 import Clock from 'clock-react';
 
 export default class QRCode extends React.Component {
@@ -8,17 +8,8 @@ export default class QRCode extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            qrcode_available: false
-        }
-    }
-
-    async componentDidMount() {
-        console.log("Rendering qr code")
-        let response = await generateQRCode();
-        if (response.status) {
-            this.setState({
-                qrcode_available: true
-            })
+            qrcode_available: true,
+            qrcode: ''
         }
     }
 
@@ -29,12 +20,12 @@ export default class QRCode extends React.Component {
                     this.state.qrcode_available ?
                         <div id={'qrcode-container'}>
                             <p id={'qrcode-description'}>Scan QR-Code to start!</p>
-                            <img src={qrcode} alt={"QRCode"} id={'qrcode-image'}/>
+                            <img src={'http://localhost:5000/public/savedQRCode/qrcode.svg'} alt={"QRCode"} id={'qrcode-image'}/>
                         </div>
                         : null
                 }
                 <div id="qr-time-container">
-                    <Clock />
+                    <Clock/>
                 </div>
             </div>
         )
