--- client/src/config.json
@@ -1,7 +1,0 @@
-{
-  "SERVER_ADDRESS": "http://127.0.0.1",
-  "SOCKET_SERVER_PORT": 5000,
-  "DJANGO_SERVER_ADDRESS": "http://localhost",
-  "DJANGO_SERVER_PORT": 8000,
-  "TRAIN_IMAGE_NUMBER": 20
-}
