+++ client/src/socketConnection.js
@@ -1,0 +1,3 @@
+import io from 'socket.io-client';
+
+export const socket = io.connect('http://localhost:5000')
