+++ config.json
@@ -1,3 +1,1 @@
-{
-  "mirror_uuid": "test-abc-123-uuid"
-}
+{"ip_host":"192.168.2.104","ip_django":"192.168.2.104:8000","django_address":"http://192.168.2.104:8000","host_address":"http://192.168.2.104:5000","uuid":"test-123-abc-uuid"}
